requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Pillow==11.2.1
pystray==0.19.5
